Finish rough planning notes
Filled out the rest of the sections in my planning notebook, picking my question and method. Does not include doing the visualizations yet. The notes on the dataset are going to have redundant elements, since not all of it is actually relevant to the question I've picked, but I'm leaving them in for my future reference in the group stage.