First draft of the dataset summary and question sections of actual report
First draft of the actual report in the first two sections. At this point it is already too long and the summary of the dataset will need to be cut down, relying on the tables to convey more information. It may also currently be going further with some analysis then appropriate for this stage.